🚧: add profile picture upload functionality and refactor EditProfileForm component
