+++ client/src/components/UploadFile.tsx
@@ -1,11 +1,34 @@
 {/* ce composant est un simple bouton qui permet d'ajouter une image pour la photo de profil */}
 import { useState } from 'react';
 
-export default function uploadButton() {
+export function uploadButton() {
     const [error, setError ] = useState('');
 
-    return (<></>);
+    return (
+        <>
+            <p>Veuillez choisir votre nouvelle photo de profil en cliquant sur le bouton.</p>
+            <input 
+                type='file' 
+                className='bg-gray-800 text-white px-3 py-1 rounded-md text-sm cursor-pointer'
+                onChange={(e) => {
+                    const file = e.target.files?.[0]; {/** déclenchement de l'évent quand un fichier est upload */}
+                    if (file) {
+                        console.log(file.name);
+                    } else {
+                        setError('No file selected');
+                    }
+                }}
+            />
+
+
+        </>
+    );
 }
+
+
+
+
+
 
 
 { /**  requete à faire coté serveur
